# docs/swagger/swagger.yaml
basePath: /
definitions:
  api.errorResponse:
    properties:
      code:
        $ref: '#/definitions/domain.Code'
      message:
        type: string
    type: object
  domain.Code:
    enum:
    - invalid_argument
    - not_found
    - already_exists
    - permission_denied
    - internal
    - unauthenticated
    type: string
    x-enum-varnames:
    - CodeInvalidArgument
    - CodeNotFound
    - CodeAlreadyExists
    - CodePermissionDenied
    - CodeInternal
    - CodeUnauthenticated
  dto.CommentReponse:
    properties:
      id:
        type: integer
      joke_id:
        type: integer
      text:
        type: string
      username:
        type: string
    type: object
  dto.CreateMyJokeRequest:
    properties:
      explanation:
        type: string
      text:
        type: string
      title:
        type: string
    type: object
...
  dto.JokeResponse:
    properties:
      explanation:
        type: string
      id:
        type: integer
      text:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
  dto.NewMyCommentRequest:
    properties:
      joke_id:
        type: integer
      text:
        type: string
    type: object
...
  dto.UpdateMyJokeExplanationRequest:
    properties:
      explanation:
        type: string
      joke_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: valeriy.tymofieiev@gmail.com
    name: API Support
    url: https://github.com/abc-valera
  description: API for Flugo social network
  license:
    name: MIT
    url: https://github.com/abc-valera
  title: Flugo
  version: "0.1"
paths:
  /comments/{joke_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      - in: query
        name: limit
        required: true
        type: integer
      - in: query
        name: offset
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - example: created_at
        in: query
        name: orderBy
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CommentReponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns comments for specified joke.
      tags:
      - comments
  /jokes:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - in: query
        name: limit
        required: true
        type: integer
      - in: query
        name: offset
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - example: created_at
        in: query
        name: orderBy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.JokeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns all jokes (limited by limit and offset) from the database.
      tags:
      - jokes
  /jokes/{joke_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: id of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JokeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns joke with specified ID from the database.
      tags:
      - jokes
  /jokes/by/{username}:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: username of required user
        in: path
        name: username
        required: true
        type: string
      - in: query
        name: limit
        required: true
        type: integer
      - in: query
        name: offset
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - example: created_at
        in: query
        name: orderBy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.JokeResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns all jokes (limited by limit and offset) by specified user.
      tags:
      - jokes
...
  /me/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: request for creating comment
        in: body
        name: comment_data
        required: true
        schema:
          $ref: '#/definitions/dto.NewMyCommentRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Creates new comment for a specified joke by current user.
      tags:
      - my comments
  /me/comments/{comment_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required comment
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Deletes specified comment by current user.
      tags:
      - my comments
...
  /me/jokes:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - in: query
        name: limit
        required: true
        type: integer
      - in: query
        name: offset
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - example: created_at
        in: query
        name: orderBy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JokeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns all jokes (limited by limit and offset) created by the current
        user.
      tags:
      - my jokes
    post:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: request for creating joke
        in: body
        name: joke_data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMyJokeRequest'
      produces:
      - text/plain
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Create new joke for current user.
      tags:
      - my jokes
  /me/jokes/{joke_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Deletes specified joke created by current user.
      tags:
      - my jokes
  /me/jokes/explanation:
    put:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: request for updating joke explanation
        in: body
        name: explanation
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMyJokeExplanationRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Updates explanation of specified joke created by current user.
      tags:
      - my jokes
...
swagger: "2.0"